refactor: move topic.tsx action into action folder
